changed variables to double type but arrays are all zeros

+++ Matrix.cpp
@@ -19,30 +19,31 @@
 const int row1 = 2, col1 = 2, row2 = 2, col2 = 2; 
 
 // method stubs
-int * multiply(int m1, int m2, int *mat1,
-    int n1, int n2, int *mat2);
+double * multiply(int m1, int m2, double *mat1,
+    int n1, int n2, double *mat2);
 
 
 int main() {
     //variable dictionary:
     int i,j,x;  //loop variables
-    int* mat1;  //first matrix (250000 x 3)
-    int* mat2;  //second matrix (3 x 7)
-    int* mat3;  //matrix resulting from mat1*mat2 (250000 x 7)
+    double* mat1;  //first matrix (250000 x 3)
+    double* mat2;  //second matrix (3 x 7)
+    double* mat3;  //matrix resulting from mat1*mat2 (250000 x 7)
 
     //standard initial output
     printf("David Dray ~ COMP 233 B ~ Spring 2020\n");
     printf("Matrix - Multiply Matrices together and "
         "compare\nthe time different thread counts "
         "take to get the answer\n");
-    mat1 = (int*)malloc(row1 * col1 * sizeof(int));
+
+    mat1 = (double*)malloc(row1 * col1 * sizeof(double));
     for (i = 0; i < row1; i++) {
         for (j = 0; j < col1; j++) {
             *(mat1 + i * col1 + j) = i+j;
         }
     }
 
-    mat2 = (int*)malloc(row2 * col2 * sizeof(int));
+    mat2 = (double*)malloc(row2 * col2 * sizeof(double));
     for (i = 0; i < row1; i++) {
         for (j = 0; j < col2; j++) {
             *(mat2 + i * col2 + j) = i+j;
@@ -71,7 +72,7 @@ int main() {
 
 
 
-    mat3 = (int*)malloc(row1 * col2 * sizeof(int));
+    mat3 = (double*)malloc(row1 * col2 * sizeof(double));
     
     mat3 = multiply(row1, col1, mat1, row2, col2, mat2);
     printf("matrix2:\n");
@@ -89,10 +90,10 @@ int main() {
 
 }
 
-int * multiply(int m1, int m2, int *mat1,
-    int n1, int n2, int *mat2) {
+double * multiply(int m1, int m2, double *mat1,
+    int n1, int n2, double *mat2) {
     int x, i, j;
-    int * res = (int*)malloc(m1 * n2 * sizeof(int));
+    double * res = (double*)malloc(m1 * n2 * sizeof(double));
     for (i = 0; i < m1; i++) {
         for (j = 0; j < n2; j++) {
